Factor classes out to separate files
Also rewrite NullLogger
